Improve initialization of some playback subjects
When player is initialized and then ready, it sends the `onReady` event which update player and start video if enabled, but playback rate and playback state were still nil as Youtube API does not send any event when ready. So i f we want to have those values before any change on those values we have to initialize the as soon as possible. At `onReady` event seems a good place to do so.